requirements.txt
# Core PyTorch dependencies
torch>=1.8.0
torchvision>=0.9.0

# Image processing
opencv-python>=4.5.0
pillow>=8.0.0
scikit-image>=0.18.0

# Data handling and processing
numpy>=1.19.0
scipy>=1.6.0
pandas>=1.2.0
pyyaml>=5.4.0

# Visualization and monitoring
matplotlib>=3.3.0
seaborn>=0.11.0
tensorboard>=2.4.0
tqdm>=4.60.0

# Metrics and evaluation
pycocotools>=2.0.2
scikit-learn>=0.24.0

# Data augmentation
albumentations>=1.0.0

# Model export
onnx>=1.8.0
onnxruntime>=1.7.0
thop>=0.0.31  # For FLOPs calculation

# Optional dependencies for specific export formats
# Uncomment as needed:
# tensorrt>=8.0.0  # For TensorRT export
# tf2onnx>=1.8.0  # For TensorFlow conversion
# openvino-dev>=2021.4.0  # For OpenVINO export
# coremltools>=5.0.0  # For CoreML export
